Steering

+++ smarty/firmware/include/gfx.h
@@ -22,7 +22,7 @@ void file_LoadImageControl(char * Datname, char * GCIName, int Mode);
 void media_VideoFrame(int X, int Y, int Framenumber);
 void putstr(char * InString);
 void WriteChars(char * charsout);
-uint32_t bus_Read(void);
+uint8_t bus_Read(void);
 void bus_In(void);
 
 #endif
